# tests/contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CriticalFlag": {
        "additionalProperties": false,
        "description": "A critical-domain concern that always forces review (AGENTS.md 10).",
        "properties": {
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "expression": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expression"
          },
          "fact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Id"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "domain",
          "reason"
        ],
        "title": "CriticalFlag",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IssueSeverity": {
        "description": "Aligned with FHIR ``issue-severity``.",
        "enum": [
          "fatal",
          "error",
          "warning",
          "information"
        ],
        "title": "IssueSeverity",
        "type": "string"
      },
      "LayerResult": {
        "additionalProperties": false,
        "description": "The outcome of one layer.",
        "properties": {
          "blocking": {
            "description": "Whether errors in this layer block the overall result.",
            "title": "Blocking",
            "type": "boolean"
          },
          "duration_ms": {
            "default": 0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "errors": {
            "default": 0,
            "title": "Errors",
            "type": "integer"
          },
          "informational": {
            "default": 0,
            "title": "Informational",
            "type": "integer"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "layer": {
            "$ref": "#/components/schemas/ValidationLayer"
          },
          "layer_number": {
            "title": "Layer Number",
            "type": "integer"
          },
          "notes": {
            "description": "Coverage caveats, e.g. which bindings could not be checked here.",
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "skipped_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skipped Reason"
          },
          "status": {
            "$ref": "#/components/schemas/LayerStatus"
          },
          "warnings": {
            "default": 0,
            "title": "Warnings",
            "type": "integer"
          }
        },
        "required": [
          "layer",
          "layer_number",
          "status",
          "blocking"
        ],
        "title": "LayerResult",
        "type": "object"
      },
      "LayerStatus": {
        "enum": [
          "passed",
          "failed",
          "skipped",
          "not_applicable"
        ],
        "title": "LayerStatus",
        "type": "string"
      },
      "ReportVersions": {
        "additionalProperties": false,
        "description": "The full version set required by principle 2.8.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fhir": {
            "title": "Fhir",
            "type": "string"
          },
          "ig": {
            "items": {
              "type": "string"
            },
            "title": "Ig",
            "type": "array"
          },
          "model": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Per-stage LLM model ids. Empty for validate-only calls.",
            "title": "Model",
            "type": "object"
          },
          "prompt_set": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Set"
          },
          "report_schema": {
            "title": "Report Schema",
            "type": "string"
          },
          "terminology": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Terminology",
            "type": "object"
          },
          "typed_models": {
            "description": "FHIR version of the fhir.resources models used for L1 (see docs/adr/0004).",
            "title": "Typed Models",
            "type": "string"
          },
          "validator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validator"
          }
        },
        "required": [
          "code",
          "report_schema",
          "fhir",
          "typed_models"
        ],
        "title": "ReportVersions",
        "type": "object"
      },
      "RoutingDecision": {
        "description": "L8 routing outcomes (AGENTS.md 10).",
        "enum": [
          "auto",
          "needs_review",
          "reject"
        ],
        "title": "RoutingDecision",
        "type": "string"
      },
...
      "ValidateRequest": {
        "additionalProperties": false,
        "description": "Body of ``POST /v1/validate``.\n\nA bare FHIR resource (anything with a top-level ``resourceType``) is also\naccepted and is rewritten into this envelope with every option defaulted, so\nan existing FHIR client can post to this endpoint without learning a wrapper\nschema.",
        "properties": {
          "layers": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationLayer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Restrict the cascade to these layers. Omit to run every layer that can run. Layers you exclude are reported as skipped, never as passed.",
            "title": "Layers"
          },
          "max_terminology_checks": {
            "default": 500,
            "description": "Cap on distinct $validate-code calls for L3. Reaching the cap is reported as a coverage note on the layer, not as a pass.",
            "maximum": 2000.0,
            "minimum": 1.0,
            "title": "Max Terminology Checks",
            "type": "integer"
          },
          "profiles": {
            "description": "Profile canonical URLs to validate against, in addition to any declared in meta.profile. A profile the validator cannot resolve fails the request rather than passing silently.",
            "examples": [
              [
                "http://hl7.org/fhir/us/core/StructureDefinition/us-core-patient"
              ]
            ],
            "items": {
              "type": "string"
            },
            "title": "Profiles",
            "type": "array"
          },
          "resource": {
            "additionalProperties": true,
            "description": "The FHIR R4 resource or Bundle to validate. Sent in the body, never in a query parameter, because it may contain PHI.",
            "title": "Resource",
            "type": "object"
          },
          "severity_overrides": {
            "additionalProperties": {
              "$ref": "#/components/schemas/IssueSeverity"
            },
            "description": "Per-rule severity overrides for L5 plausibility, keyed by rule id.",
            "title": "Severity Overrides",
            "type": "object"
          }
        },
        "required": [
          "resource"
        ],
        "title": "ValidateRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationIssue": {
        "additionalProperties": false,
        "description": "One issue from one layer.",
        "properties": {
          "code": {
            "description": "FHIR issue-type code, or a rule id for L5.",
            "title": "Code",
            "type": "string"
          },
          "column": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column"
          },
          "expression": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "FHIRPath location, e.g. Bundle.entry[4].resource.category.",
            "title": "Expression"
          },
          "layer": {
            "$ref": "#/components/schemas/ValidationLayer"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "machine_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "A fhirbridge error code, when the issue maps to one.",
            "title": "Machine Code"
          },
          "message": {
            "description": "Human-readable detail. May quote element values.",
            "title": "Message",
            "type": "string"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Identifier of the L5 plausibility rule that fired.",
            "title": "Rule Id"
          },
          "severity": {
            "$ref": "#/components/schemas/IssueSeverity"
          }
        },
        "required": [
          "layer",
          "severity",
          "code",
          "message"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
      "ValidationLayer": {
        "description": "The eight cascade layers, in execution order.",
        "enum": [
          "structural",
          "profile",
          "terminology",
          "invariants",
          "plausibility",
          "fidelity",
          "coverage",
          "routing"
        ],
        "title": "ValidationLayer",
        "type": "string"
      },
      "ValidationReport": {
        "additionalProperties": false,
        "description": "The response body of ``POST /v1/validate``.",
        "properties": {
          "conformant": {
            "description": "True when no blocking issue was raised by any layer that ran.",
            "title": "Conformant",
            "type": "boolean"
          },
          "conversion_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversion Id"
          },
          "critical_flags": {
            "items": {
              "$ref": "#/components/schemas/CriticalFlag"
            },
            "title": "Critical Flags",
            "type": "array"
          },
          "duration_ms": {
            "default": 0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/LayerResult"
            },
            "title": "Layers",
            "type": "array"
          },
          "nondeterminism_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Nondeterminism Reasons",
            "type": "array"
          },
          "nondeterminism_risk": {
            "default": false,
            "title": "Nondeterminism Risk",
            "type": "boolean"
          },
          "omissions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Omissions",
            "type": "array"
          },
          "profiles": {
            "items": {
              "type": "string"
            },
            "title": "Profiles",
            "type": "array"
          },
          "resource_count": {
            "description": "1, or the number of Bundle entries.",
            "title": "Resource Count",
            "type": "integer"
          },
          "resource_type": {
            "title": "Resource Type",
            "type": "string"
          },
          "scores": {
            "$ref": "#/components/schemas/ValidationScores"
          },
          "status": {
            "$ref": "#/components/schemas/RoutingDecision"
          },
          "validated_at": {
            "format": "date-time",
            "title": "Validated At",
            "type": "string"
          },
          "versions": {
            "$ref": "#/components/schemas/ReportVersions"
          }
        },
        "required": [
          "status",
          "conformant",
          "resource_type",
          "resource_count",
          "layers",
          "versions"
        ],
        "title": "ValidationReport",
        "type": "object"
      },
      "ValidationScores": {
        "additionalProperties": false,
        "description": "Scores. ``None`` means \"not measured\", never \"perfect\".",
        "properties": {
          "conformance": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Fraction of submitted resources with zero blocking issues in L1-L4. Defined in docs/evaluation.md.",
            "title": "Conformance"
          },
          "coverage": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "L7 coverage, i.e. 1 - omission_rate.",
            "title": "Coverage"
          },
          "fidelity": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "L6 semantic fidelity. Requires source spans and an LLM.",
            "title": "Fidelity"
          },
          "mean_confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Mean calibrated fact confidence. Requires a conversion.",
            "title": "Mean Confidence"
          }
        },
        "title": "ValidationScores",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "A scoped API key (`fhirb_<prefix>_<secret>`) or an OAuth2 client-credentials access token. API keys are for sandbox use; production deployments should use OAuth2 or SMART Backend Services.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "fhirbridge",
      "url": "https://github.com/fhirbridge/fhirbridge"
    },
    "description": "Convert unstructured clinical source material into validated, provenance-tagged\nFHIR R4 resources, with a mandatory human-review stage before anything is\nfinalized.\n\n**What this is not:** not a medical device, not clinical decision support, and not\na FHIR repository. Nothing is inferred beyond what the source text asserts, and no\nendpoint writes to an external FHIR server without an explicit, separate call.\n\n**Fail-closed behaviour.** When the FHIR validator or the terminology server is\nunavailable, requests fail with `503` rather than returning an unverified result.\nAn unvalidated code and an unvalidated bundle are worse than no answer.\n\n**Errors.** Clinical and FHIR-semantic failures return an `OperationOutcome`\nwhose `issue.details.coding` carries a stable machine code from\n`https://fhirbridge.org/CodeSystem/errors`. Platform failures return\n`{\"error\": {\"code\", \"message\", \"trace_id\", \"details\"}}`. Fetch the full code\nlist from `GET /v1/error-codes`.\n\n**No PHI in URLs.** Every endpoint that accepts clinical content accepts it in a\nrequest body. There is no `GET` endpoint taking clinical text as a query\nparameter, and none will be added.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "summary": "Narrative-to-FHIR conversion with a verification harness.",
...
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/fhir/R4/$validate": {
      "post": {
        "description": "``POST /fhir/R4/$validate``.\n\nAccepts either a bare resource or a ``Parameters`` resource carrying\n``resource`` and optional ``profile`` parameters, which is how FHIR clients\nnormally invoke the operation.",
        "operationId": "post_fhir_validate",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/fhir+json": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Validator or terminology server unavailable; failing closed."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Validate a resource (runs the same cascade as POST /v1/validate)",
        "tags": [
          "fhir-facade"
        ]
      }
    },
...
    "/v1/validate": {
      "post": {
        "operationId": "post_validate",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "vital_sign": {
                  "value": {
                    "profiles": [
                      "http://hl7.org/fhir/us/core/StructureDefinition/us-core-vital-signs"
                    ],
                    "resource": {
                      "code": {
                        "coding": [
                          {
                            "code": "8867-4",
                            "display": "Heart rate",
                            "system": "http://loinc.org"
                          }
                        ]
                      },
                      "resourceType": "Observation",
                      "status": "preliminary",
                      "subject": {
                        "reference": "Patient/example"
                      },
                      "valueQuantity": {
                        "code": "/min",
                        "system": "http://unitsofmeasure.org",
                        "unit": "beats/minute",
                        "value": 72
                      }
                    }
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "The payload is not a structurally valid FHIR resource."
          },
          "422": {
            "description": "A requested profile is not loaded in the validator sidecar."
          },
          "503": {
            "description": "The validator or terminology server is unavailable. The request fails closed rather than returning an unverified report."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Validate a FHIR resource or Bundle through the L1-L5 cascade",
        "tags": [
          "validation"
        ]
      }
    },
    "/v1/validate/outcome": {
      "post": {
        "description": "The same cascade, rendered as an ``OperationOutcome``.\n\nOffered because FHIR tooling expects a validation call to answer with an\nOperationOutcome. The report shape carries strictly more information, so\nprefer ``POST /v1/validate`` when the client is ours.",
        "operationId": "post_validate_outcome",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/fhir+json": {}
            },
            "description": "An OperationOutcome carrying one issue per finding."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Validate and return a FHIR OperationOutcome instead of a report",
        "tags": [
          "validation"
        ]
      }
    },
...
  },
...
}